// src/odata-client/sfo-data-service/CustomPayTypeAssignment.d.ts
export declare class CustomPayTypeAssignment extends Entity implements CustomPayTypeAssignmentType {
    /**
     * Technical entity name for CustomPayTypeAssignment.
     */
    static _entityName: string;
    /**
     * @deprecated Since v1.0.1 Use [[_defaultServicePath]] instead.
     * Technical service name for CustomPayTypeAssignment.
     */
    static _serviceName: string;
    /**
     * Default url path for the according service.
     */
    static _defaultServicePath: string;
    /**
     * Customized Pay Type_External Code.
     * Maximum length: 128.
     */
    customPayTypeExternalCode: string;
    /**
     * Country/Region.
     * Maximum length: 128.
     * @nullable
     */
    country?: string;
    /**
     * createdBy.
     * Maximum length: 100.
     * @nullable
     */
    createdBy?: string;
    /**
     * createdDate.
     * @nullable
     */
    createdDateTime?: Moment;
    /**
     * externalCode.
     */
    externalCode: BigNumber;
    /**
     * lastModifiedBy.
     * Maximum length: 100.
     * @nullable
     */
    lastModifiedBy?: string;
    /**
     * lastModifiedDate.
     * @nullable
     */
    lastModifiedDateTime?: Moment;
    /**
     * Record Status.
     * Maximum length: 255.
     * @nullable
     */
    mdfSystemRecordStatus?: string;
    /**
     * One-to-many navigation property to the [[Country]] entity.
     */
    countryNav: Country[];
    /**
     * One-to-one navigation property to the [[User]] entity.
     */
    createdByNav: User;
    /**
     * One-to-one navigation property to the [[User]] entity.
     */
    lastModifiedByNav: User;
    /**
     * One-to-one navigation property to the [[MdfEnumValue]] entity.
     */
    mdfSystemRecordStatusNav: MdfEnumValue;
    /**
     * Returns an entity builder to construct instances `CustomPayTypeAssignment`.
     * @returns A builder that constructs instances of entity type `CustomPayTypeAssignment`.
     */
    static builder(): EntityBuilderType<CustomPayTypeAssignment, CustomPayTypeAssignmentTypeForceMandatory>;
    /**
     * Returns a request builder to construct requests for operations on the `CustomPayTypeAssignment` entity type.
     * @returns A `CustomPayTypeAssignment` request builder.
     */
    static requestBuilder(): CustomPayTypeAssignmentRequestBuilder;
    /**
     * Returns a selectable object that allows the selection of custom field in a get request for the entity `CustomPayTypeAssignment`.
     * @param fieldName Name of the custom field to select
     * @returns A builder that constructs instances of entity type `CustomPayTypeAssignment`.
     */
    static customField(fieldName: string): CustomField<CustomPayTypeAssignment>;
    /**
     * Overwrites the default toJSON method so that all instance variables as well as all custom fields of the entity are returned.
     * @returns An object containing all instance variables + custom fields.
     */
    toJSON(): {
        [key: string]: any;
    };
}
import { Country, CountryType } from './Country';
import { User, UserType } from './User';
import { MdfEnumValue, MdfEnumValueType } from './MdfEnumValue';
export interface CustomPayTypeAssignmentType {
    customPayTypeExternalCode: string;
    country?: string;
    createdBy?: string;
    createdDateTime?: Moment;
    externalCode: BigNumber;
    lastModifiedBy?: string;
    lastModifiedDateTime?: Moment;
    mdfSystemRecordStatus?: string;
    countryNav: CountryType[];
    createdByNav: UserType;
    lastModifiedByNav: UserType;
    mdfSystemRecordStatusNav: MdfEnumValueType;
}
export interface CustomPayTypeAssignmentTypeForceMandatory {
    customPayTypeExternalCode: string;
    country: string;
    createdBy: string;
    createdDateTime: Moment;
    externalCode: BigNumber;
    lastModifiedBy: string;
    lastModifiedDateTime: Moment;
    mdfSystemRecordStatus: string;
    countryNav: CountryType[];
    createdByNav: UserType;
    lastModifiedByNav: UserType;
    mdfSystemRecordStatusNav: MdfEnumValueType;
}
export declare namespace CustomPayTypeAssignment {
    /**
     * Static representation of the [[customPayTypeExternalCode]] property for query construction.
     * Use to reference this property in query operations such as 'select' in the fluent request API.
     */
    const CUSTOM_PAY_TYPE_EXTERNAL_CODE: StringField<CustomPayTypeAssignment>;
    /**
     * Static representation of the [[country]] property for query construction.
     * Use to reference this property in query operations such as 'select' in the fluent request API.
     */
    const COUNTRY: StringField<CustomPayTypeAssignment>;
    /**
     * Static representation of the [[createdBy]] property for query construction.
     * Use to reference this property in query operations such as 'select' in the fluent request API.
     */
    const CREATED_BY: StringField<CustomPayTypeAssignment>;
    /**
     * Static representation of the [[createdDateTime]] property for query construction.
     * Use to reference this property in query operations such as 'select' in the fluent request API.
     */
    const CREATED_DATE_TIME: DateField<CustomPayTypeAssignment>;
    /**
     * Static representation of the [[externalCode]] property for query construction.
     * Use to reference this property in query operations such as 'select' in the fluent request API.
     */
    const EXTERNAL_CODE: BigNumberField<CustomPayTypeAssignment>;
    /**
     * Static representation of the [[lastModifiedBy]] property for query construction.
     * Use to reference this property in query operations such as 'select' in the fluent request API.
     */
    const LAST_MODIFIED_BY: StringField<CustomPayTypeAssignment>;
    /**
     * Static representation of the [[lastModifiedDateTime]] property for query construction.
     * Use to reference this property in query operations such as 'select' in the fluent request API.
     */
    const LAST_MODIFIED_DATE_TIME: DateField<CustomPayTypeAssignment>;
    /**
     * Static representation of the [[mdfSystemRecordStatus]] property for query construction.
     * Use to reference this property in query operations such as 'select' in the fluent request API.
     */
    const MDF_SYSTEM_RECORD_STATUS: StringField<CustomPayTypeAssignment>;
    /**
     * Static representation of the one-to-many navigation property [[countryNav]] for query construction.
     * Use to reference this property in query operations such as 'select' in the fluent request API.
     */
    const COUNTRY_NAV: Link<CustomPayTypeAssignment, Country>;
    /**
     * Static representation of the one-to-one navigation property [[createdByNav]] for query construction.
     * Use to reference this property in query operations such as 'select' in the fluent request API.
     */
    const CREATED_BY_NAV: OneToOneLink<CustomPayTypeAssignment, User>;
    /**
     * Static representation of the one-to-one navigation property [[lastModifiedByNav]] for query construction.
     * Use to reference this property in query operations such as 'select' in the fluent request API.
     */
    const LAST_MODIFIED_BY_NAV: OneToOneLink<CustomPayTypeAssignment, User>;
    /**
     * Static representation of the one-to-one navigation property [[mdfSystemRecordStatusNav]] for query construction.
     * Use to reference this property in query operations such as 'select' in the fluent request API.
     */
    const MDF_SYSTEM_RECORD_STATUS_NAV: OneToOneLink<CustomPayTypeAssignment, MdfEnumValue>;
    /**
     * All fields of the CustomPayTypeAssignment entity.
     */
    const _allFields: Array<StringField<CustomPayTypeAssignment> | DateField<CustomPayTypeAssignment> | BigNumberField<CustomPayTypeAssignment> | Link<CustomPayTypeAssignment, Country> | OneToOneLink<CustomPayTypeAssignment, User> | OneToOneLink<CustomPayTypeAssignment, MdfEnumValue>>;
    /**
     * All fields selector.
     */
    const ALL_FIELDS: AllFields<CustomPayTypeAssignment>;
    /**
     * All key fields of the CustomPayTypeAssignment entity.
     */
    const _keyFields: Array<Selectable<CustomPayTypeAssignment>>;
    /**
     * Mapping of all key field names to the respective static field property CustomPayTypeAssignment.
     */
    const _keys: {
        [keys: string]: Selectable<CustomPayTypeAssignment>;
    };
}